+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+- Fixes bug where month and year `Datepicker` calendars were rendering vertically rather than horizontally.
+
 ## 2.1.3 (December 7, 2023)
 
 ### Adds